+++ MovPhantSlave.X/config_slave.h
@@ -35,7 +35,7 @@ int32_t* makeWideVelPulseWaveform(int32_t amplitude, uint16_t numValues);
 int32_t* makeZeroWaveform(uint16_t numValues);
 int32_t* makeConstWaveform(int32_t value, uint16_t numValues);
 void allocateArbitraryWaveform(uint16_t nPts);
-void setWaveformValue(int32_t value, uint16_t whichWaveform);
+void setWaveformValue(uint16_t value, uint16_t whichWaveform);
